+++ src/main/java/com/google/cloud/spanner/pgadapter/statements/BackendConnection.java
@@ -14,6 +14,7 @@
 
 package com.google.cloud.spanner.pgadapter.statements;
 
+import static com.google.cloud.spanner.pgadapter.error.PGExceptionFactory.toPGException;
 import static com.google.cloud.spanner.pgadapter.statements.SimpleParser.isCommand;
 import static com.google.cloud.spanner.pgadapter.wireprotocol.QueryMessage.COPY;
 
@@ -29,7 +30,6 @@ import com.google.cloud.spanner.Partition;
 import com.google.cloud.spanner.PartitionOptions;
 import com.google.cloud.spanner.ReadContext.QueryAnalyzeMode;
 import com.google.cloud.spanner.ResultSet;
-import com.google.cloud.spanner.ResultSets;
 import com.google.cloud.spanner.Spanner;
 import com.google.cloud.spanner.SpannerBatchUpdateException;
 import com.google.cloud.spanner.SpannerException;
@@ -58,12 +58,14 @@ import com.google.cloud.spanner.pgadapter.statements.DdlExecutor.NotExecuted;
 import com.google.cloud.spanner.pgadapter.statements.SessionStatementParser.SessionStatement;
 import com.google.cloud.spanner.pgadapter.statements.SimpleParser.TableOrIndexName;
 import com.google.cloud.spanner.pgadapter.statements.local.LocalStatement;
+import com.google.cloud.spanner.pgadapter.utils.ClientAutoDetector.WellKnownClient;
 import com.google.cloud.spanner.pgadapter.utils.CopyDataReceiver;
 import com.google.cloud.spanner.pgadapter.utils.MutationWriter;
 import com.google.cloud.spanner.pgadapter.wireoutput.ReadyResponse;
 import com.google.cloud.spanner.pgadapter.wireoutput.ReadyResponse.Status;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
+import com.google.common.base.Suppliers;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMap.Builder;
@@ -86,6 +88,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.function.Function;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 
@@ -135,7 +138,7 @@ public class BackendConnection {
   }
 
   static <T> PGException setAndReturn(SettableFuture<T> future, Throwable throwable) {
-    PGException pgException = PGExceptionFactory.toPGException(throwable);
+    PGException pgException = toPGException(throwable);
     future.setException(pgException);
     return pgException;
   }
@@ -207,9 +210,10 @@ public class BackendConnection {
         //  block always ends with a ROLLBACK, PGAdapter should skip the entire execution of that
         //  block.
         SessionStatement sessionStatement = getSessionManagementStatement(parsedStatement);
-        if (!localStatements.isEmpty() && localStatements.containsKey(statement.getSql())) {
+        if (!localStatements.get().isEmpty()
+            && localStatements.get().containsKey(statement.getSql())) {
           result.set(
-              Objects.requireNonNull(localStatements.get(statement.getSql()))
+              Objects.requireNonNull(localStatements.get().get(statement.getSql()))
                   .execute(BackendConnection.this));
         } else if (sessionStatement != null) {
           result.set(sessionStatement.execute(sessionState));
@@ -252,7 +256,7 @@ public class BackendConnection {
           // Potentially replace pg_catalog table references with common table expressions.
           updatedStatement =
               sessionState.isReplacePgCatalogTables()
-                  ? pgCatalog.replacePgCatalogTables(statement)
+                  ? pgCatalog.get().replacePgCatalogTables(statement)
                   : statement;
           updatedStatement = bindStatement(updatedStatement);
           result.set(analyzeOrExecute(updatedStatement));
@@ -371,7 +375,7 @@ public class BackendConnection {
     }
   }
 
-  private static final int MAX_PARTITIONS =
+  public static final int MAX_PARTITIONS =
       Math.max(16, 2 * Runtime.getRuntime().availableProcessors());
 
   private final class CopyOut extends BufferedStatement<StatementResult> {
@@ -400,9 +404,14 @@ public class BackendConnection {
               // No need for the extra complexity of a partitioned query.
               result.set(spannerConnection.execute(statement));
             } else {
+              // Get the metadata of the query, so we can include that in the result.
+              ResultSet metadataResultSet =
+                  spannerConnection.analyzeQuery(statement, QueryAnalyzeMode.PLAN);
               result.set(
                   new PartitionQueryResult(
-                      batchReadOnlyTransaction.getBatchTransactionId(), partitions));
+                      batchReadOnlyTransaction.getBatchTransactionId(),
+                      partitions,
+                      metadataResultSet));
             }
           } catch (SpannerException spannerException) {
             // The query might not be suitable for partitioning. Just try with a normal query.
@@ -476,6 +485,84 @@ public class BackendConnection {
     }
   }
 
+  private final class Vacuum extends BufferedStatement<StatementResult> {
+    final VacuumStatement vacuumStatement;
+
+    Vacuum(VacuumStatement vacuumStatement) {
+      super(vacuumStatement.parsedStatement, vacuumStatement.originalStatement);
+      this.vacuumStatement = vacuumStatement;
+    }
+
+    @Override
+    void execute() {
+      try {
+        checkConnectionState();
+        if (spannerConnection.isInTransaction()) {
+          throw PGExceptionFactory.newPGException(
+              "VACUUM cannot run inside a transaction block", SQLState.ActiveSqlTransaction);
+        }
+        for (TableOrIndexName table : vacuumStatement.getTables()) {
+          ImmutableList<TableOrIndexName> columns = vacuumStatement.getTableColumns(table);
+          Statement statement;
+          if (columns == null || columns.isEmpty()) {
+            statement = Statement.of("select * from " + table + " limit 1");
+          } else {
+            statement =
+                Statement.of(
+                    "select "
+                        + columns.stream().map(c -> c.name).collect(Collectors.joining(","))
+                        + " from "
+                        + table
+                        + " limit 1");
+          }
+          // Just analyze the query to ensure the table and column names are valid.
+          spannerConnection.analyzeQuery(statement, QueryAnalyzeMode.PLAN);
+        }
+        result.set(NO_RESULT);
+      } catch (Exception exception) {
+        result.setException(toPGException(exception));
+        throw exception;
+      }
+    }
+  }
+
+  private final class Truncate extends BufferedStatement<StatementResult> {
+    final TruncateStatement truncateStatement;
+
+    Truncate(TruncateStatement truncateStatement) {
+      super(truncateStatement.parsedStatement, truncateStatement.originalStatement);
+      this.truncateStatement = truncateStatement;
+    }
+
+    @Override
+    void execute() {
+      try {
+        checkConnectionState();
+        if (spannerConnection.isDdlBatchActive()) {
+          throw PGExceptionFactory.newPGException("Cannot execute TRUNCATE in a DDL batch");
+        }
+        if (spannerConnection.isDmlBatchActive()) {
+          throw PGExceptionFactory.newPGException("Cannot execute TRUNCATE in a DML batch");
+        }
+        if (spannerConnection.isInTransaction()
+            || spannerConnection.getAutocommitDmlMode() == AutocommitDmlMode.TRANSACTIONAL) {
+          spannerConnection.executeBatchUpdate(truncateStatement.getDeleteStatements());
+        } else {
+          // BatchDML is not supported for PDML, so we have to loop over the statements.
+          // We do not execute the statements in parallel, as there might be dependencies between
+          // the tables that are being truncated.
+          for (Statement statement : truncateStatement.getDeleteStatements()) {
+            spannerConnection.executeUpdate(statement);
+          }
+        }
+        result.set(NO_RESULT);
+      } catch (Exception exception) {
+        result.setException(exception);
+        throw exception;
+      }
+    }
+  }
+
   private static final ImmutableMap<String, LocalStatement> EMPTY_LOCAL_STATEMENTS =
       ImmutableMap.of();
   private static final StatementResult NO_RESULT = new NoResult();
@@ -483,8 +570,8 @@ public class BackendConnection {
   private static final Statement ROLLBACK = Statement.of("ROLLBACK");
 
   private final SessionState sessionState;
-  private final PgCatalog pgCatalog;
-  private final ImmutableMap<String, LocalStatement> localStatements;
+  private final Supplier<PgCatalog> pgCatalog;
+  private final Supplier<ImmutableMap<String, LocalStatement>> localStatements;
   private ConnectionState connectionState = ConnectionState.IDLE;
   private TransactionMode transactionMode = TransactionMode.IMPLICIT;
   private final String currentSchema = "public";
@@ -497,24 +584,31 @@ public class BackendConnection {
   BackendConnection(
       DatabaseId databaseId,
       Connection spannerConnection,
+      Supplier<WellKnownClient> wellKnownClient,
       OptionsMetadata optionsMetadata,
-      ImmutableList<LocalStatement> localStatements) {
+      Supplier<ImmutableList<LocalStatement>> localStatements) {
     this.sessionState = new SessionState(optionsMetadata);
-    this.pgCatalog = new PgCatalog(this.sessionState);
+    this.pgCatalog =
+        Suppliers.memoize(
+            () -> new PgCatalog(BackendConnection.this.sessionState, wellKnownClient.get()));
     this.spannerConnection = spannerConnection;
     this.databaseId = databaseId;
     this.ddlExecutor = new DdlExecutor(databaseId, this);
-    if (localStatements.isEmpty()) {
-      this.localStatements = EMPTY_LOCAL_STATEMENTS;
-    } else {
-      Builder<String, LocalStatement> builder = ImmutableMap.builder();
-      for (LocalStatement localStatement : localStatements) {
-        for (String sql : localStatement.getSql()) {
-          builder.put(new SimpleImmutableEntry<>(sql, localStatement));
-        }
-      }
-      this.localStatements = builder.build();
-    }
+    this.localStatements =
+        Suppliers.memoize(
+            () -> {
+              if (localStatements.get().isEmpty()) {
+                return EMPTY_LOCAL_STATEMENTS;
+              } else {
+                Builder<String, LocalStatement> builder = ImmutableMap.builder();
+                for (LocalStatement localStatement : localStatements.get()) {
+                  for (String sql : localStatement.getSql()) {
+                    builder.put(new SimpleImmutableEntry<>(sql, localStatement));
+                  }
+                }
+                return builder.build();
+              }
+            });
   }
 
   /** Returns the current connection state. */
@@ -536,7 +630,8 @@ public class BackendConnection {
     return execute.result;
   }
 
-  public Future<StatementResult> analyze(ParsedStatement parsedStatement, Statement statement) {
+  public ListenableFuture<StatementResult> analyze(
+      ParsedStatement parsedStatement, Statement statement) {
     Execute execute = new Execute(parsedStatement, statement, Function.identity(), true);
     bufferedStatements.add(execute);
     return execute.result;
@@ -563,6 +658,18 @@ public class BackendConnection {
     CopyOut copyOut = new CopyOut(parsedStatement, statement);
     bufferedStatements.add(copyOut);
     return copyOut.result;
+  }
+
+  public Future<StatementResult> execute(VacuumStatement vacuumStatement) {
+    Vacuum vacuum = new Vacuum(vacuumStatement);
+    bufferedStatements.add(vacuum);
+    return vacuum.result;
+  }
+
+  public Future<StatementResult> execute(TruncateStatement truncateStatement) {
+    Truncate truncate = new Truncate(truncateStatement);
+    bufferedStatements.add(truncate);
+    return truncate.result;
   }
 
   /** Flushes the buffered statements to Spanner. */
@@ -1170,10 +1277,15 @@ public class BackendConnection {
   public static final class PartitionQueryResult implements StatementResult {
     private final BatchTransactionId batchTransactionId;
     private final List<Partition> partitions;
+    private final ResultSet metadataResultSet;
 
-    public PartitionQueryResult(BatchTransactionId batchTransactionId, List<Partition> partitions) {
+    public PartitionQueryResult(
+        BatchTransactionId batchTransactionId,
+        List<Partition> partitions,
+        ResultSet metadataResultSet) {
       this.batchTransactionId = batchTransactionId;
       this.partitions = partitions;
+      this.metadataResultSet = metadataResultSet;
     }
 
     public BatchTransactionId getBatchTransactionId() {
@@ -1182,6 +1294,10 @@ public class BackendConnection {
 
     public List<Partition> getPartitions() {
       return partitions;
+    }
+
+    public ResultSet getMetadataResultSet() {
+      return metadataResultSet;
     }
 
     @Override
@@ -1196,7 +1312,7 @@ public class BackendConnection {
 
     @Override
     public ResultSet getResultSet() {
-      return ResultSets.forRows(
+      return ClientSideResultSet.forRows(
           Type.struct(StructField.of("partition", Type.bytes())),
           partitions.stream()
               .map(